+++ database/migrations/2024_03_30_100450_add_attribute_id_to_order_log_component_attributes_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! Schema::hasTable('order_log_component_attributes')) {
+            return;
+        }
+
         Schema::table('order_log_component_attributes', function (Blueprint $table) {
             $table->integer('attribute_id')->after('attribute_key')->nullable();
         });
@@ -21,6 +25,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! Schema::hasTable('order_log_component_attributes')) {
+            return;
+        }
+
         Schema::table('order_log_component_attributes', function (Blueprint $table) {
             $table->dropColumn('attribute_id');
         });
